refactor: fixed comments
