+++ src/components/order/constants.tsx
@@ -2,16 +2,18 @@ import { createColumnHelper } from '@tanstack/react-table';
 
 import { Box } from '@mui/material';
 
-import { statusColors, type IStatusColors } from '@/styles/order-status-colors';
+import { type IOrder, type IStatusColors } from './types';
 
-interface IOrder {
-  orderNumber: string;
-  orderDate: string;
-  customer: string;
-  amount: string;
-  deliveryDate: string;
-  status: string;
-}
+const statusColors: IStatusColors = {
+  Open: { background: '#FFF6ED', text: '#FFA347' },
+  Ordered: { background: '#FFEEF6', text: '#EC4899' },
+  'In Progress': { background: '#EFF8FF', text: '#175CD3' },
+  'Ready to use': { background: '#F3F3FF', text: '#6366F1' },
+  'Ready to ship': { background: '#ECFDF3', text: '#12B76A' },
+  Sent: { background: '#CEFDE2', text: '#12B76A' },
+  Delivered: { background: '#F8FAFC', text: '#475569' },
+  Canceled: { background: '#FEF3F2', text: '#F04438' },
+};
 
 function createOrderData(
   orderNumber: string,
