+++ src/pages/index.tsx
@@ -5,8 +5,6 @@ import { Box, CircularProgress } from '@mui/material';
 
 import { RootLayout } from '../layouts/root-layout';
 
-import AddOrderPage from './add-order-page/add-order-page';
-
 const HomePage = lazy(() => {
   return import('@/pages/home-page/home-page');
 });
@@ -25,6 +23,10 @@ const StationPage = lazy(() => {
 
 const OrderPage = lazy(() => {
   return import('@/pages/order-page/order-page');
+});
+
+const AddOrderPage = lazy(() => {
+  return import('@/pages/add-order-page/add-order-page');
 });
 
 function PageLoader(): React.ReactNode {
@@ -52,9 +54,9 @@ export function App(): React.ReactNode {
             <Route path="home" element={<HomePage />} />
             <Route path="material" element={<MaterialPage />} />
             <Route path="material/add-material" element={<AddMaterialPage />} />
-            <Route path="order/add-order" element={<AddOrderPage />} />
             <Route path="station" element={<StationPage />} />
             <Route path="order" element={<OrderPage />} />
+            <Route path="order/add-order" element={<AddOrderPage />} />
           </Route>
         </Routes>
       </Suspense>
